project/requirements/common.txt
#requirements/common.txt
Django==1.6
Fabric==1.8.2
MySQL-python==1.2.5
South==0.8.1
lxml==3.3.3
six==1.6.1
uWSGI==2.0.2
django-celery==3.1.10
django-countries==2.1.2
selenium==2.41.0
cloudinary==1.0.16
beautifulsoup4==4.3.2
mongoengine==0.8.7